@@ -1,1 +1,2 @@
-export { default as middleware } from './middleware/index';
+import './install-fetch.js';
+export { default as middleware } from './middleware/index.js';
