@@ -1,6 +1,7 @@
 import fs from 'fs';
 import path from 'path';
-import mime from 'mime/lite';
+import * as mime from 'mime/lite.js';
+const { getType } = mime;
 import { Handler, SapperRequest, SapperResponse, build_dir, dev, manifest } from '@sapper/internal/manifest-server';
 import { get_server_route_handler } from './get_server_route_handler';
 import { get_page_handler } from './get_page_handler';
@@ -110,12 +111,12 @@ export function serve({ prefix, pathname, cache_control }: {
 		? (file: string) => fs.readFileSync(path.join(build_dir, file))
 		: (file: string) => (cache.has(file) ? cache : cache.set(file, fs.readFileSync(path.join(build_dir, file)))).get(file);
 
-	return (req: SapperRequest, res: SapperResponse, next: () => void) => {
+	return async (req: SapperRequest, res: SapperResponse, next: () => void) => {
 		if (filter(req)) {
-			const type = mime.getType(req.path);
+			const type = getType(req.path);
 
 			try {
-				const file = path.posix.normalize(decodeURIComponent(req.path));
+				const file = path.posix.normalize(path.join('.', decodeURIComponent(req.path)));
 				const data = read(file);
 
 				res.setHeader('Content-Type', type);
