@@ -1,4 +1,4 @@
-export { dev } from './api/dev';
-export { build } from './api/build';
-export { export } from './api/export';
-export { find_page } from './api/find_page';
+export { dev } from './api/dev.js';
+export { build } from './api/build.js';
+export { export as export } from './api/export.js';
+export { find_page } from './api/find_page.js';
