@@ -4,16 +4,17 @@ import * as http from 'http';
 import * as child_process from 'child_process';
 import * as ports from 'port-authority';
 import { EventEmitter } from 'events';
-import { create_manifest_data, create_app, create_compilers, create_serviceworker_manifest } from '../core';
-import { Compiler, Compilers } from '../core/create_compilers';
-import { CompileResult } from '../core/create_compilers/interfaces';
-import Deferred from './utils/Deferred';
-import validate_bundler from './utils/validate_bundler';
-import { copy_shimport } from './utils/copy_shimport';
-import { ManifestData, FatalEvent, ErrorEvent, ReadyEvent, InvalidEvent } from '../interfaces';
-import { noop } from './utils/noop';
-import { copy_runtime } from './utils/copy_runtime';
-import { rimraf, mkdirp } from './utils/fs_utils';
+import CheapWatch from 'cheap-watch';
+import { create_manifest_data, create_app, create_compilers, create_serviceworker_manifest } from '../core.js';
+import { Compiler, Compilers } from '../core/create_compilers/index.js';
+import { CompileResult } from '../core/create_compilers/interfaces.js';
+import Deferred from './utils/Deferred.js';
+import validate_bundler from './utils/validate_bundler.js';
+import { copy_shimport } from './utils/copy_shimport.js';
+import { ManifestData, FatalEvent, ErrorEvent, ReadyEvent, InvalidEvent } from '../interfaces.js';
+import { noop } from './utils/noop.js';
+import { copy_runtime } from './utils/copy_runtime.js';
+import { rimraf, mkdirp } from './utils/fs_utils.js';
 
 type Opts = {
 	cwd?: string;
@@ -141,11 +142,11 @@ class Watcher extends EventEmitter {
 
 		rimraf(output);
 		mkdirp(output);
-		copy_runtime(output);
+		await copy_runtime(output);
 
 		rimraf(dest);
 		mkdirp(`${dest}/client`);
-		if (this.bundler === 'rollup') copy_shimport(dest);
+		if (this.bundler === 'rollup') await copy_shimport(dest);
 
 		if (!this.dev_port) this.dev_port = await ports.find(10000);
 
@@ -289,7 +290,7 @@ class Watcher extends EventEmitter {
 							this.emit('stderr', chunk);
 						});
 
-						this.proc.on('message', message => {
+						this.proc.on('message', (message: ExportMessage) => {
 							if (message.__sapper__ && message.event === 'basepath') {
 								this.emit('basepath', {
 									basepath: message.basepath
@@ -493,17 +494,14 @@ function watch_dir(
 	let watch: any;
 	let closed = false;
 
-	import('cheap-watch').then(({ default: CheapWatch }) => {
-		if (closed) return;
+	if (closed) return;
+	watch = new CheapWatch({ dir, filter, debounce: 50 }) as CheapWatch;
 
-		watch = new CheapWatch({ dir, filter, debounce: 50 });
+	watch.on('+', callback);
 
-		watch.on('+', callback);
+	watch.on('-', callback);
 
-		watch.on('-', callback);
-
-		watch.init();
-	});
+	watch.init();
 
 	return {
 		close: () => {
@@ -511,4 +509,10 @@ function watch_dir(
 			closed = true;
 		}
 	};
+}
+
+interface ExportMessage {
+	__sapper__: boolean;
+	event: string;
+	basepath: string;
 }
