@@ -1,10 +1,18 @@
-import * as fs from 'fs';
+import { writeFileSync } from 'fs';
+import { readFile } from 'fs/promises';
 import * as path from 'path';
 import sade from 'sade';
 import colors from 'kleur';
-import * as pkg from '../package.json';
-import { elapsed, repeat, left_pad, format_milliseconds } from './utils';
-import { InvalidEvent, ErrorEvent, FatalEvent, BuildEvent, ReadyEvent } from './interfaces';
+import { resolve } from 'import-meta-resolve';
+import * as pb from 'pretty-bytes';
+import { elapsed, repeat, left_pad, format_milliseconds } from './utils.js';
+import { InvalidEvent, ErrorEvent, FatalEvent, BuildEvent, ReadyEvent } from './interfaces.js';
+
+const pkg = JSON.parse(
+	(await readFile(
+		new URL(await resolve('../package.json', import.meta.url)).pathname
+	)).toString()
+)
 
 const prog = sade('sapper').version(pkg.version);
 
@@ -47,7 +55,7 @@ prog.command('dev')
 		'build-dir': string;
 		ext: string;
 	}) => {
-		const { dev } = await import('./api/dev');
+		const { dev } = await import('./api/dev.js');
 
 		try {
 			const watcher = dev({
@@ -173,7 +181,7 @@ prog.command('build [dest]')
 
 			const launcher = path.resolve(dest, 'index.js');
 
-			fs.writeFileSync(launcher, `
+			writeFileSync(launcher, `
 				// generated by sapper build at ${new Date().toISOString()}
 				process.env.NODE_ENV = process.env.NODE_ENV || 'production';
 				process.env.PORT = process.env.PORT || ${opts.port || 3000};
@@ -231,8 +239,7 @@ prog.command('export [dest]')
 				console.error(`\n> Built in ${elapsed(start)}`);
 			}
 
-			const { export: _export } = await import('./api/export');
-			const { default: pb } = await import('pretty-bytes');
+			const { export: _export } = await import('./api/export.js');
 
 			await _export({
 				cwd: opts.cwd,
@@ -281,7 +288,7 @@ async function _build(
 	dest: string,
 	ext: string
 ) {
-	const { build } = await import('./api/build');
+	const { build } = await import('./api/build.js');
 
 	await build({
 		bundler,
