@@ -1,3 +1,3 @@
-export * from './core/create_app';
-export { default as create_compilers } from './core/create_compilers/index';
-export { default as create_manifest_data } from './core/create_manifest_data';
+export * from './core/create_app.js';
+export { default as create_compilers } from './core/create_compilers/index.js';
+export { default as create_manifest_data } from './core/create_manifest_data.js';
