@@ -1,7 +1,7 @@
 import * as fs from 'fs';
 import * as path from 'path';
-import { posixify, stringify, walk, write_if_changed } from '../utils';
-import { Page, PageComponent, ManifestData } from '../interfaces';
+import { posixify, stringify, walk, write_if_changed } from '../utils.js';
+import { Page, PageComponent, ManifestData } from '../interfaces.js';
 
 export function create_app({
 	bundler,
@@ -33,8 +33,8 @@ export function create_app({
 
 	const app = generate_app(manifest_data, path_to_routes);
 
-	write_if_changed(`${output}/internal/manifest-client.mjs`, client_manifest);
-	write_if_changed(`${output}/internal/manifest-server.mjs`, server_manifest);
+	write_if_changed(`${output}/internal/manifest-client.js`, client_manifest);
+	write_if_changed(`${output}/internal/manifest-server.js`, server_manifest);
 	write_if_changed(`${output}/internal/App.svelte`, app);
 }
 
@@ -148,7 +148,7 @@ function generate_client_manifest(
 		export const routes = ${routes};
 
 		${dev ? `if (typeof window !== 'undefined') {
-			import(${stringify(posixify(path.resolve(__dirname, '../sapper-dev-client.js')))}).then(client => {
+			import(${stringify(posixify(path.resolve(path.dirname(new URL(import.meta.url).pathname), '../sapper-dev-client.js')))}).then(client => {
 				client.connect(${dev_port});
 			});
 		}` : ''}
@@ -263,7 +263,7 @@ function generate_app(manifest_data: ManifestData, path_to_routes: string) {
 		<!-- This file is generated by Sapper — do not edit it! -->
 		<script>
 			import { setContext, afterUpdate } from 'svelte';
-			import { CONTEXT_KEY } from './shared';
+			import { CONTEXT_KEY } from './shared.js';
 			import Layout from '${get_file(path_to_routes, manifest_data.root)}';
 			import Error from '${get_file(path_to_routes, manifest_data.error)}';
 
