@@ -1,8 +1,8 @@
 import * as path from 'path';
-import color from 'kleur';
-import relative from 'require-relative';
-import { dependenciesForTree, DependencyTreeOptions } from 'rollup-dependency-tree';
-import css_chunks from 'rollup-plugin-css-chunks';
+import colors from 'kleur';
+import css_chunks_o from 'rollup-plugin-css-chunks';
+const { default: css_chunks } = css_chunks_o as any;
+import relative from './require-relative.js';
 import {
 	NormalizedOutputOptions,
 	OutputBundle,
@@ -12,8 +12,10 @@ import {
 	RenderedChunk,
 	RollupError
 } from 'rollup';
-import { CompileResult } from './interfaces';
-import RollupResult from './RollupResult';
+import { dependenciesForTree, DependencyTreeOptions } from './rollup-dependency-tree.js';
+import { CompileResult } from './interfaces.js';
+import RollupResult from './RollupResult.js';
+import { get_config_extname } from './config_extname.js';
 
 const stderr = console.error.bind(console);
 const INJECT_STYLES_NAME = 'inject_styles';
@@ -21,7 +23,7 @@ const INJECT_STYLES_ID = 'inject_styles.js';
 
 let rollup: any;
 
-function printTimings(timings: {[event: string]: [number, number, number]}) {
+function printTimings(timings:{ [event: string]: [number, number, number] }) {
 	for (const [key, info] of Object.entries(timings)) {
 		console.info(`${key} took ${info[0].toFixed(0)}ms`);
 	}
@@ -117,6 +119,7 @@ export default class RollupCompiler {
 	css_files: Record<string, string>;
 	dependencies: Record<string, string[]>;
 	routes: string;
+	esm: boolean;
 
 	constructor(config: any, routes: string) {
 		this._ = this.get_config(config);
@@ -317,7 +320,7 @@ export default class RollupCompiler {
 		return mod;
 	}
 
-	oninvalid(cb: (filename: string) => void) {
+	async oninvalid(cb: (filename: string) => void) {
 		this._oninvalid = cb;
 	}
 
@@ -353,7 +356,7 @@ export default class RollupCompiler {
 			this.warnings = [];
 			this.errors = [];
 			this._oninvalid(id);
-		});
+		})
 
 		watcher.on('event', (event: any) => {
 			switch (event.code) {
@@ -395,14 +398,23 @@ export default class RollupCompiler {
 				default:
 					console.log(`Unexpected event ${event.code}`);
 			}
-		});
+		})
 	}
 
 	static async load_config(cwd: string) {
 		if (!rollup) rollup = relative('rollup', cwd);
 
-		const input = path.resolve(cwd, 'rollup.config.js');
-
+		let input: string, format: string, esm: boolean;
+		const config_extname = await get_config_extname(cwd);
+		if (config_extname === '.cjs') {
+			input = path.resolve(cwd, 'rollup.config.js');
+			format = 'cjs';
+			esm = false;
+		} else if (config_extname === '.mjs') {
+			input = path.resolve(cwd, 'rollup.config.mjs');
+			format = 'es';
+			esm = true;
+		}
 		const bundle = await rollup.rollup({
 			input,
 			inlineDynamicImports: true,
@@ -415,26 +427,30 @@ export default class RollupCompiler {
 			output: [{ code }]
 		} = await bundle.generate({
 			exports: 'named',
-			format: 'cjs'
+			format
 		});
 
-		// temporarily override require
-		const defaultLoader = require.extensions['.js'];
-		require.extensions['.js'] = (module: any, filename: string) => {
-			if (filename === input) {
-				module._compile(code, filename);
-			} else {
-				defaultLoader(module, filename);
+		if (!esm) {
+			// temporarily override require
+			const defaultLoader = require.extensions['.js'];
+			require.extensions['.js'] = (module: any, filename: string) => {
+				if (filename === input) {
+					module._compile(code, filename);
+				} else {
+					defaultLoader(module, filename);
+				}
 			}
-		};
+		}
 
-		const config: any = require(input).default; // eslint-disable-line
-		delete require.cache[input];
-
-		return config;
+		if (esm) {
+			return (await import(input)).default;
+		} else {
+			const config = require(input).default;
+			delete require.cache[input];
+			return config;
+		}
 	}
 }
-
 
 // copied from https://github.com/rollup/rollup/blob/master/cli/logging.ts
 // and updated so that it will compile here
@@ -444,13 +460,13 @@ export function handleError(err: RollupError, recover = false) {
 	if (err.name) description = `${err.name}: ${description}`;
 	const message =
 		(err.plugin
-			? `(plugin ${(err).plugin}) ${description}`
-			: description) || err;
+		 ? `(plugin ${(err).plugin}) ${description}`
+		 : description) || err;
 
-	stderr(color.bold().red(`[!] ${color.bold(message.toString())}`));
+	stderr(colors.bold().red(`[!] ${colors.bold(message.toString())}`));
 
 	if (err.url) {
-		stderr(color.cyan(err.url));
+		stderr(colors.cyan(err.url));
 	}
 
 	if (err.loc) {
@@ -460,11 +476,11 @@ export function handleError(err: RollupError, recover = false) {
 	}
 
 	if (err.frame) {
-		stderr(color.dim(err.frame));
+		stderr(colors.dim(err.frame));
 	}
 
 	if (err.stack) {
-		stderr(color.dim(err.stack));
+		stderr(colors.dim(err.stack));
 	}
 
 	stderr('');
