@@ -1,7 +1,8 @@
 import * as path from 'path';
-import RollupCompiler from './RollupCompiler';
-import { WebpackCompiler } from './WebpackCompiler';
-import { set_dev, set_src, set_dest } from '../../config/env';
+import RollupCompiler from './RollupCompiler.js';
+import { WebpackCompiler } from './WebpackCompiler.js';
+import { set_dev, set_src, set_dest } from '../../config/env.js';
+import { get_config_extname } from './config_extname.js'
 
 export type Compiler = RollupCompiler | WebpackCompiler;
 
@@ -25,7 +26,7 @@ export default async function create_compilers(
 
 	if (bundler === 'rollup') {
 		const config = await RollupCompiler.load_config(cwd);
-		validate_config(config, 'rollup');
+		await validate_config(cwd, config, 'rollup');
 
 		normalize_rollup_config(config.client);
 		normalize_rollup_config(config.server);
@@ -42,8 +43,8 @@ export default async function create_compilers(
 	}
 
 	if (bundler === 'webpack') {
-		const config = require(path.resolve(cwd, 'webpack.config.js')); // eslint-disable-line
-		validate_config(config, 'webpack');
+		const config = await import(path.resolve(cwd, 'webpack.config.js'))
+		await validate_config(cwd, config, 'webpack');
 
 		return {
 			client: new WebpackCompiler(config.client),
@@ -56,9 +57,10 @@ export default async function create_compilers(
 	throw new Error(`Invalid bundler option '${bundler}'`);
 }
 
-function validate_config(config: any, bundler: 'rollup' | 'webpack') {
+async function validate_config(cwd, config: any, bundler: 'rollup' | 'webpack') {
 	if (!config.client || !config.server) {
-		throw new Error(`${bundler}.config.js must export a { client, server, serviceworker? } object`);
+		const extname = await get_config_extname(cwd);
+		throw new Error(`${bundler}.config.${extname} must export a { client, server, serviceworker? } object`);
 	}
 }
 
